refactor: consolidate service examples into grouped files
