+++ example_user_test.go
@@ -26,6 +26,10 @@ var (
 	doerRecentlyViewedListProjects = newMockDoer(fixture.RecentlyViewed.ProjectListJSON)
 	doerRecentlyViewedListWikis    = newMockDoer(fixture.RecentlyViewed.WikiListJSON)
 	doerRecentlyViewedAddWiki      = newMockDoer(fixture.RecentlyViewed.WikiSingleJSON)
+
+	// UserStarService
+	doerUserStarCount = newMockDoer(fixture.Star.CountJSON)
+	doerUserStarList  = newMockDoer(fixture.Star.ListJSON)
 )
 
 func ExampleUserService_All() {
@@ -193,4 +197,30 @@ func ExampleUserRecentlyViewedService_AddWiki() {
 	fmt.Printf("Name: %s\n", wiki.Name)
 	// Output:
 	// Name: Home
+}
+
+func ExampleUserStarService_Count() {
+	c, _ := backlog.NewClient(
+		"https://example.backlog.com",
+		"token",
+		backlog.WithDoer(doerUserStarCount),
+	)
+
+	count, _ := c.User.Star.Count(context.Background(), 1)
+	fmt.Printf("Count: %d\n", count)
+	// Output:
+	// Count: 42
+}
+
+func ExampleUserStarService_List() {
+	c, _ := backlog.NewClient(
+		"https://example.backlog.com",
+		"token",
+		backlog.WithDoer(doerUserStarList),
+	)
+
+	stars, _ := c.User.Star.List(context.Background(), 1)
+	fmt.Printf("ID: %d, Title: %s\n", stars[0].ID, stars[0].Title)
+	// Output:
+	// ID: 10, Title: [TEST-1] first issue
 }
